UnicornProd/requirements.txt
attrs==19.3.0
aws-cdk.assets==1.33.0
aws-cdk.aws-applicationautoscaling==1.33.0
aws-cdk.aws-autoscaling-common==1.33.0
aws-cdk.aws-cloudformation==1.33.0
aws-cdk.aws-cloudwatch==1.33.0
aws-cdk.aws-dynamodb==1.33.0
aws-cdk.aws-ec2==1.33.0
aws-cdk.aws-events==1.33.0
aws-cdk.aws-iam==1.33.0
aws-cdk.aws-kms==1.33.0
aws-cdk.aws-lambda==1.33.0
aws-cdk.aws-logs==1.33.0
aws-cdk.aws-s3==1.33.0
aws-cdk.aws-s3-assets==1.33.0
aws-cdk.aws-sns==1.33.0
aws-cdk.aws-sqs==1.33.0
aws-cdk.aws-ssm==1.33.0
aws-cdk.cloud-assembly-schema==1.33.0
aws-cdk.core==1.33.0
aws-cdk.custom-resources==1.33.0
aws-cdk.cx-api==1.33.0
aws-cdk.region-info==1.33.0
cattrs==1.0.0
constructs==3.0.1
jsii==1.3.0
publication==0.0.3
python-dateutil==2.8.1
six==1.14.0
typing-extensions==3.7.4.2
-e git+https://github.com/mateja82/unicorn@b381724e628fd8ddfbae9c408321fa1e181c2efb#egg=unicorn_prod&subdirectory=UnicornProd
